requirements.txt
streamlit
ultralytics
opencv-python-headless
pillow
numpy
torch
torchvision
wandb
requests
pyyaml
annotated-types
huggingface_hub
